Cured some Motif bugs
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@1844 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/motif/checklst.cpp
@@ -60,13 +60,13 @@ wxCheckListBox::wxCheckListBox(wxWindow *parent, wxWindowID id,
 // check items
 // -----------
 
-bool wxCheckListBox::IsChecked(uint uiIndex) const
+bool wxCheckListBox::IsChecked(size_t uiIndex) const
 {
     // TODO
     return FALSE;
 }
 
-void wxCheckListBox::Check(uint uiIndex, bool bCheck)
+void wxCheckListBox::Check(size_t uiIndex, bool bCheck)
 {
     // TODO
 }
